+++ lib/Braintree/CreditCardVerification.php
@@ -24,6 +24,7 @@ namespace Braintree;
  * @property-read string $processorResponseText
  * @property-read string $processorResponseType
  * @property-read \Braintree\RiskData|null $riskData
+ * @property-read \Braintree\ThreeDSecureInfo|null $threeDSecureInfo
  */
 class CreditCardVerification extends Result\CreditCardVerification
 {