+++ lib/Braintree/TransactionGateway.php
@@ -525,7 +525,11 @@ class TransactionGateway
                    'expected transaction id to be set'
                    );
         }
-        if (!preg_match('/^[0-9a-z]+$/', $id)) {
+        
+        // NEXT_MAJOR_VERSION - none of the other sdks validate the format
+        // of the ID. In the next major version, we can remove this check 
+        // and have the gateway return a 404 error instead
+        if (!preg_match('/^[0-9a-z_]+$/', $id)) {
             throw new InvalidArgumentException(
                     $id . ' is an invalid transaction id.'
                     );