+++ tests/integration/CreditCardTest.php
@@ -788,6 +788,32 @@ class CreditCardTest extends Setup
         Braintree\CreditCard::fromNonce($nonce);
     }
 
+    public function testCreate_fromThreeDSecureNonce()
+    {
+        $customer = Braintree\Customer::createNoValidate();
+        $http = new HttpClientApi(Braintree\Configuration::$global);
+        $nonce = Braintree\Test\Nonces::$threeDSecureVisaFullAuthenticationNonce;
+
+        $result = Braintree\CreditCard::create([
+            'customerId' => $customer->id,
+            'paymentMethodNonce' => $nonce,
+            'options' => [
+                'verifyCard' => 'true',
+            ]
+        ]);
+
+        $threeDSecureInfo = $result->creditCard->verification->threeDSecureInfo;
+        $this->assertTrue($threeDSecureInfo->liabilityShiftPossible);
+        $this->assertTrue($threeDSecureInfo->liabilityShifted);
+        $this->assertEquals("Y", $threeDSecureInfo->enrolled);
+        $this->assertEquals("authenticate_successful", $threeDSecureInfo->status);
+        $this->assertEquals("xid_value", $threeDSecureInfo->xid);
+        $this->assertEquals("cavv_value", $threeDSecureInfo->cavv);
+        $this->assertEquals("05", $threeDSecureInfo->eciFlag);
+        $this->assertEquals(null, $threeDSecureInfo->dsTransactionId);
+        $this->assertEquals("1.0.2", $threeDSecureInfo->threeDSecureVersion);
+    }
+
     public function testFromNonce_ReturnsErrorWhenNonceIsConsumed()
     {
         $customer = Braintree\Customer::createNoValidate();